// bridge/typechain-types/StakingSVGMock.ts
/* Autogenerated file. Do not edit manually. */
/* tslint:disable */
/* eslint-disable */
import { FunctionFragment, Result } from "@ethersproject/abi";
import { Listener, Provider } from "@ethersproject/providers";
import {
  BaseContract,
  BigNumber,
  BytesLike,
  CallOverrides,
  PopulatedTransaction,
  Signer,
  utils,
} from "ethers";
import { OnEvent, TypedEvent, TypedEventFilter, TypedListener } from "./common";

export declare namespace StakingSVG {
  export type StakingSVGParamsStruct = {
    shares: string;
    freeAfter: string;
    withdrawFreeAfter: string;
    accumulatorEth: string;
    accumulatorToken: string;
  };

  export type StakingSVGParamsStructOutput = [
    string,
    string,
    string,
    string,
    string
  ] & {
    shares: string;
    freeAfter: string;
    withdrawFreeAfter: string;
    accumulatorEth: string;
    accumulatorToken: string;
  };
}

export interface StakingSVGMockInterface extends utils.Interface {
  contractName: "StakingSVGMock";
  functions: {
    "generateSVG((string,string,string,string,string))": FunctionFragment;
  };

  encodeFunctionData(
    functionFragment: "generateSVG",
    values: [StakingSVG.StakingSVGParamsStruct]
  ): string;

  decodeFunctionResult(
    functionFragment: "generateSVG",
    data: BytesLike
  ): Result;

  events: {};
}

export interface StakingSVGMock extends BaseContract {
  contractName: "StakingSVGMock";
  connect(signerOrProvider: Signer | Provider | string): this;
  attach(addressOrName: string): this;
  deployed(): Promise<this>;

  interface: StakingSVGMockInterface;

  queryFilter<TEvent extends TypedEvent>(
    event: TypedEventFilter<TEvent>,
    fromBlockOrBlockhash?: string | number | undefined,
    toBlock?: string | number | undefined
  ): Promise<Array<TEvent>>;

  listeners<TEvent extends TypedEvent>(
    eventFilter?: TypedEventFilter<TEvent>
  ): Array<TypedListener<TEvent>>;
  listeners(eventName?: string): Array<Listener>;
  removeAllListeners<TEvent extends TypedEvent>(
    eventFilter: TypedEventFilter<TEvent>
  ): this;
  removeAllListeners(eventName?: string): this;
  off: OnEvent<this>;
  on: OnEvent<this>;
  once: OnEvent<this>;
  removeListener: OnEvent<this>;

  functions: {
    generateSVG(
      svgParams: StakingSVG.StakingSVGParamsStruct,
      overrides?: CallOverrides
    ): Promise<[string] & { svg: string }>;
  };

  generateSVG(
    svgParams: StakingSVG.StakingSVGParamsStruct,
    overrides?: CallOverrides
  ): Promise<string>;

  callStatic: {
    generateSVG(
      svgParams: StakingSVG.StakingSVGParamsStruct,
      overrides?: CallOverrides
    ): Promise<string>;
  };

  filters: {};

  estimateGas: {
    generateSVG(
      svgParams: StakingSVG.StakingSVGParamsStruct,
      overrides?: CallOverrides
    ): Promise<BigNumber>;
  };

  populateTransaction: {
    generateSVG(
      svgParams: StakingSVG.StakingSVGParamsStruct,
      overrides?: CallOverrides
    ): Promise<PopulatedTransaction>;
  };
}